Pulls in the netCDF files into the long/lat projection, rather than an Albers projection.
Outputs leaflet maps instead of the ggplot maps.  This is an extensive change.  Currently only continuous output is plotted.

There is an issue with the plotting of the last SD plot.  I'm not sure why.  It seems to have something to do with the last palette.

The UI allows two or four panels on a conditional (if the second taxon is 'None' or not).