For some reason Quentin's files won't delete properly.
Changed the `file` assignment in the `downloadHandler` on the `server.R` file.

Added a small script to take the ncdf output & zip it with the data description.

Added the `data_description`.  The file is up for editing currently on Google Drive.

Modified the `README.md` text.  Moved Jody & Michelle to Contributors & Chris & Simon to Development.  Modified the program function block to add links to pre-prints & publications.

Updated the version history.